Refactored files and organized some imports
