+++ src/main/webapp/resources/frontend/teamworks/src/components/settings/department/DepartmentSettings.js
@@ -1,14 +1,9 @@
-import SettingGroup from "./SettingGroup";
-import EditableField from "./EditableField";
-import Button from "../buttons/Button";
-import AddUserForm from "../forms/AddUserForm";
-import UserList from "./DepartmentMemberList";
-import AddElementForm from "../forms/AddElementForm";
-import DepartmentsContainer from "./DepartmentsContainer";
-import AddDepartmentForm from "../forms/AddDepartmentForm";
 import { useEffect, useState } from "react";
-import DepartmentApiUtils from "../../utils/api/DepartmentApiUtils";
-import Spinner from "../spinner/Spinner";
+import DepartmentApiUtils from "../../../utils/api/DepartmentApiUtils";
+import AddDepartmentForm from "../../forms/AddDepartmentForm";
+import Spinner from "../../spinner/Spinner";
+import SettingGroup from "../SettingGroup";
+import DepartmentsContainer from "./DepartmentsContainer";
 
 const DepartmentSettings = () =>
 {
