+++ src/main/webapp/resources/frontend/teamworks/src/components/settings/users/UserSettings.js
@@ -1,7 +1,5 @@
-import SettingGroup from "./SettingGroup";
-import EditableField from "./EditableField";
-import Button from "../buttons/Button";
-import AddUserForm from "../forms/AddUserForm";
+import SettingGroup from "../SettingGroup";
+import AddUserForm from "../../forms/AddUserForm";
 //import UserList from "./UserList";
 
 const UserSettings = () => {
